Updated last advanced using long double

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,14 +5,14 @@
 */
 int main(void)
 {
-	double n1 = 1, n2 = 2, r;
+	long double n1 = 1, n2 = 2, r;
 	int i;
 
 	printf("1, 2, ");
 	for (i = 0; i < 96; i++)
 	{
 		r = n1 + n2;
-		printf("%.0f", r);
+		printf("%.0Lf", r);
 		n1 = n2;
 		n2 = r;
 		if (i != 95)
